+++ activerecord/test/cases/test_fixtures_test.rb
@@ -21,6 +21,17 @@ class TestFixturesTest < ActiveRecord::TestCase
     assert_equal "foobar", @klass.use_transactional_tests
   end
 
+  def test_inclusion_runs_active_record_fixtures_load_hook
+    ActiveSupport.on_load(:active_record_fixtures) do
+      self.fixture_paths << "test/fixtures"
+    end
+    klass = Class.new
+
+    klass.include(ActiveRecord::TestFixtures)
+
+    assert_includes klass.fixture_paths, "test/fixtures"
+  end
+
   unless in_memory_db?
     def test_doesnt_rely_on_active_support_test_case_specific_methods
       tmp_dir = Dir.mktmpdir